Split out common test out of e2e_logging.py
